updated project files
old git did not work so the project has been moved to this file.
So far these systems have been developed:
- Movement
- Weapon holstering and pistol shooting
- Sitting down in the flying seat
- flying controls

Also the scene for the main ship has been made, we will need to build the ship now and make sure that the player is attached to the ship when not flying and make sure the flying still works.